Padding added to arrays causes float32's to be incorrectly cast to float64s in the case where a record batch only contains one row.
This issue also applies to 32-bit integers miscast to 64-bit integers,
however on little endian machines the cast + truncation results in a
correct value.

In the float32 case, the mis-read 64-bit float is denormal and so when
it's converted to 32-bit it becomes zero.

This only occurs in the case where there's one (4-byte) element.