Fix datafusion error(s) when query results are empty.
There are two issues at play. The first is in when a comparison
operation has one (or both) sides returning a NullArray. This type
cannot downcast to the appropriate typed array and so results in
a panic. The fix is to check to see if one side of the expression
is all nulls, and if so, propagate that as the result of the
expresssion.

The second fix is that, if the query results in no records, the
last step of the query (sorting the results) was assuming that there
was at least one record batch.

If no record batches were generated as part of the query, an error
would be propagated (however it would not panic). But an error is
not an acceptable result, rather the result should be an empty
result set.

This change creates the empty result set if there are no generated
record batches.